Trait halo2::arithmetic::BaseExt

pub trait BaseExt: Field + Ord + ConstantTimeEq {
    const MODULUS: &'static str;

    fn from_bytes_wide(bytes: &[u8; 64]) -> Self;

    fn write<W>(&self, writer: &mut W) -> Result<(), Error>
    where
        W: Write;

    fn read<R>(reader: &mut R) -> Result<Self, Error>
    where
        R: Read;

    fn rand() -> Self { ... }

    fn ct_is_zero(&self) -> Choice { ... }

    fn pow(&self, by: &[u64; 4]) -> Self { ... }

    fn batch_invert(v: &mut [Self]) -> Self { ... }
}

This trait is a common interface for dealing with elements of a finite field.

Associated Constants

const MODULUS: &'static str

Modulus of the field written as a string for display purposes

Required methods

fn from_bytes_wide(bytes: &[u8; 64]) -> Self

Obtains a field element that is congruent to the provided little endian byte representation of an integer.

fn write<W>(&self, writer: &mut W) -> Result<(), Error> where
W: Write,

Writes this element in its normalized, little endian form into a buffer.

fn read<R>(reader: &mut R) -> Result<Self, Error> where
R: Read,

Reads a normalized, little endian represented field element from a buffer.

Provided methods

fn rand() -> Self

This computes a random element of the field using system randomness.

fn ct_is_zero(&self) -> Choice

Returns whether or not this element is zero.

fn pow(&self, by: &[u64; 4]) -> Self

Exponentiates self by by, where by is a little-endian order integer exponent.

fn batch_invert(v: &mut [Self]) -> Self

Performs a batch inversion using Montgomery’s trick, returns the product of every inverse. Zero inputs are ignored.