Trait halo2::arithmetic::BatchInvert

pub trait BatchInvert<F: Field> {
    fn batch_invert(self) -> F;
}

Extension trait for iterators over mutable field elements which allows those field elements to be inverted in a batch.

Required methods

fn batch_invert(self) -> F

Consume this iterator and invert each field element (when nonzero), returning the inverse of all nonzero field elements. Zero elements are left as zero.