Trait halo2::arithmetic::CurveAffine

pub trait CurveAffine: PrimeCurveAffine<Scalar = Self::ScalarExt, Curve = Self::CurveExt> + Default + Add<Self, Output = Self::Curve> + Sub<Self, Output = Self::Curve> + ConditionallySelectable + ConstantTimeEq + From<Self::Curve> {
    type ScalarExt: FieldExt;
    type Base: FieldExt;
    type CurveExt: CurveExt;
    fn coordinates(&self) -> CtOption<Coordinates<Self>>;

    fn from_xy(x: Self::Base, y: Self::Base) -> CtOption<Self>;

    fn is_on_curve(&self) -> Choice;

    fn a() -> Self::Base;

    fn b() -> Self::Base;

    fn read<R>(reader: &mut R) -> Result<Self, Error>
    where
        R: Read,
    { ... }

    fn write<W>(&self, writer: &mut W) -> Result<(), Error>
    where
        W: Write,
    { ... }
}

This trait is the affine counterpart to Curve and is used for serialization, storage in memory, and inspection of $x$ and $y$ coordinates.

Associated Types

type ScalarExt: FieldExt

The scalar field of this elliptic curve.

type Base: FieldExt

The base field over which this elliptic curve is constructed.

type CurveExt: CurveExt

The projective form of the curve

Required methods

fn coordinates(&self) -> CtOption<Coordinates<Self>>

Gets the coordinates of this point.

Returns None if this is the identity.

fn from_xy(x: Self::Base, y: Self::Base) -> CtOption<Self>

Obtains a point given $(x, y)$, failing if it is not on the curve.

fn is_on_curve(&self) -> Choice

Returns whether or not this element is on the curve; should always be true unless an “unchecked” API was used.

fn a() -> Self::Base

Returns the curve constant $a$.

fn b() -> Self::Base

Returns the curve constant $b$.

Provided methods

fn read<R>(reader: &mut R) -> Result<Self, Error> where
R: Read,

Reads a compressed element from the buffer and attempts to parse it using from_bytes.

fn write<W>(&self, writer: &mut W) -> Result<(), Error> where
W: Write,

Writes an element in compressed form to the buffer.