Trait halo2::arithmetic::CurveExt

pub trait CurveExt: PrimeCurve<Affine = Self::AffineExt, Scalar = Self::ScalarExt> + Group + Default + ConditionallySelectable + ConstantTimeEq + From<Self::Affine> + Group<Scalar = Self::Scalar> {
    type ScalarExt: FieldExt;
    type Base: BaseExt;
    type AffineExt: CurveAffine + Mul<Self::ScalarExt>;

    const CURVE_ID: &'static str;

    fn jacobian_coordinates(&self) -> (Self::Base, Self::Base, Self::Base);

    fn is_on_curve(&self) -> Choice;

    fn b() -> Self::Base;

    fn new_jacobian(
        x: Self::Base, 
        y: Self::Base, 
        z: Self::Base
    ) -> CtOption<Self>;
}

This trait is a common interface for dealing with elements of an elliptic curve group in a “projective” form, where that arithmetic is usually more efficient.

Associated Types

type ScalarExt: FieldExt

The scalar field of this elliptic curve.

type Base: BaseExt

The base field over which this elliptic curve is constructed.

type AffineExt: CurveAffine + Mul<Self::ScalarExt>

The affine version of the curve

Associated Constants

const CURVE_ID: &'static str

CURVE_ID used for hash-to-curve.

Required methods

fn jacobian_coordinates(&self) -> (Self::Base, Self::Base, Self::Base)

Apply the curve endomorphism by multiplying the x-coordinate by an element of multiplicative order 3. Return the Jacobian coordinates of this point.

fn is_on_curve(&self) -> Choice

Returns whether or not this element is on the curve; should always be true unless an “unchecked” API was used.

fn b() -> Self::Base

Returns the curve constant a. Returns the curve constant b.

fn new_jacobian(x: Self::Base, y: Self::Base, z: Self::Base) -> CtOption<Self>

Obtains a point given Jacobian coordinates $X : Y : Z$, failing if the coordinates are not on the curve.