Trait halo2::arithmetic::Engine

pub trait Engine: 'static + Clone {
    type Scalar: FieldExt;
    type G1: PrimeCurve + From<Self::G1Affine> + GroupOps<Self::G1Affine, Self::G1> + GroupOpsOwned<Self::G1Affine, Self::G1> + ScalarMul<Self::Scalar, Self::G1> + ScalarMulOwned<Self::Scalar, Self::G1> + Group;
    type G1Affine: PairingCurveAffine + From<Self::G1> + Mul<Self::Scalar> + for<'a> Mul<&'a Self::Scalar>;
    type G2: PrimeCurve + From<Self::G2Affine> + GroupOps<Self::G2Affine, Self::G2> + GroupOpsOwned<Self::G2Affine, Self::G2> + ScalarMul<Self::Scalar, Self::G2> + ScalarMulOwned<Self::Scalar, Self::G2>;
    type G2Affine: PairingCurveAffine + From<Self::G2> + Mul<Self::Scalar> + for<'a> Mul<&'a Self::Scalar>;
    type Gt: Group + ScalarMul<Self::Scalar, Self::Gt> + ScalarMulOwned<Self::Scalar, Self::Gt>;
    fn pairing(p: &Self::G1Affine, q: &Self::G2Affine) -> Self::Gt;
}

Associated Types

type Scalar: FieldExt

This is the scalar field of the engine’s groups.

type G1: PrimeCurve + From<Self::G1Affine> + GroupOps<Self::G1Affine, Self::G1> + GroupOpsOwned<Self::G1Affine, Self::G1> + ScalarMul<Self::Scalar, Self::G1> + ScalarMulOwned<Self::Scalar, Self::G1> + Group

The projective representation of an element in G1.

type G1Affine: PairingCurveAffine + From<Self::G1> + Mul<Self::Scalar> + for<'a> Mul<&'a Self::Scalar>

The affine representation of an element in G1.

type G2: PrimeCurve + From<Self::G2Affine> + GroupOps<Self::G2Affine, Self::G2> + GroupOpsOwned<Self::G2Affine, Self::G2> + ScalarMul<Self::Scalar, Self::G2> + ScalarMulOwned<Self::Scalar, Self::G2>

The projective representation of an element in G2.

type G2Affine: PairingCurveAffine + From<Self::G2> + Mul<Self::Scalar> + for<'a> Mul<&'a Self::Scalar>

The affine representation of an element in G2.

type Gt: Group + ScalarMul<Self::Scalar, Self::Gt> + ScalarMulOwned<Self::Scalar, Self::Gt>

The extension field that hosts the target group of the pairing.

Required methods

fn pairing(p: &Self::G1Affine, q: &Self::G2Affine) -> Self::Gt

Invoke the pairing function G1 x G2 -> Gt without the use of precomputation and other optimizations.