Trait halo2::arithmetic::FieldExt

pub trait FieldExt: PrimeField + From<bool> + Ord + Group<Scalar = Self> {
    const MODULUS: &'static str;
    const ROOT_OF_UNITY: Self;
    const ROOT_OF_UNITY_INV: Self;
    const T_MINUS1_OVER2: [u64; 4];
    const DELTA: Self;
    const TWO_INV: Self;
    const RESCUE_ALPHA: u64;
    const RESCUE_INVALPHA: [u64; 4];
    const ZETA: Self;

    fn sqrt_ratio(num: &Self, div: &Self) -> (Choice, Self);

    fn from_u64(v: u64) -> Self;

    fn from_u128(v: u128) -> Self;

    fn to_bytes(&self) -> [u8; 32];

    fn from_bytes(bytes: &[u8; 32]) -> CtOption<Self>;

    fn from_bytes_wide(bytes: &[u8; 64]) -> Self;

    fn get_lower_128(&self) -> u128;

    fn get_lower_32(&self) -> u32;

    fn sqrt_alt(&self) -> (Choice, Self) { ... }

    fn rand() -> Self { ... }

    fn write<W>(&self, writer: &mut W) -> Result<(), Error>
    where
        W: Write,
    { ... }

    fn read<R>(reader: &mut R) -> Result<Self, Error>
    where
        R: Read,
    { ... }

    fn pow(&self, by: &[u64; 4]) -> Self { ... }

    fn pow_by_t_minus1_over2(&self) -> Self { ... }

}

This trait is a common interface for dealing with elements of a finite field.

Associated Constants

const MODULUS: &'static str

Modulus of the field written as a string for display purposes

const ROOT_OF_UNITY: Self

Generator of the $2^S$ multiplicative subgroup

const ROOT_OF_UNITY_INV: Self

Inverse of ROOT_OF_UNITY

const T_MINUS1_OVER2: [u64; 4]

The value $(T-1)/2$ such that $2^S \cdot T = p - 1$ with $T$ odd.

const DELTA: Self

Generator of the $t-order$ multiplicative subgroup

const TWO_INV: Self

Inverse of $2$ in the field.

const RESCUE_ALPHA: u64

Ideally the smallest prime $\alpha$ such that gcd($p - 1$, $\alpha$) = $1$

const RESCUE_INVALPHA: [u64; 4]

$RESCUE_INVALPHA \cdot RESCUE_ALPHA = 1 \mod p - 1$ such that (a^RESCUE_ALPHA)^RESCUE_INVALPHA = a.

const ZETA: Self

Element of multiplicative order $3$.

Required methods

fn sqrt_ratio(num: &Self, div: &Self) -> (Choice, Self)

Computes:

(true, sqrt(num/div)), if num and div are nonzero and num/div is a square in the field;

(true, 0), if num is zero;

(false, 0), if num is nonzero and div is zero;

(false, sqrt(ROOT_OF_UNITY * num/div)), if num and div are nonzero and num/div is a nonsquare in the field;

where ROOT_OF_UNITY is a generator of the order 2^n subgroup (and therefore a nonsquare).

The choice of root from sqrt is unspecified.

fn from_u64(v: u64) -> Self

Obtains a field element congruent to the integer v.

fn from_u128(v: u128) -> Self

Obtains a field element congruent to the integer v.

fn to_bytes(&self) -> [u8; 32]

Converts this field element to its normalized, little endian byte representation.

fn from_bytes(bytes: &[u8; 32]) -> CtOption<Self>

Attempts to obtain a field element from its normalized, little endian byte representation.

fn from_bytes_wide(bytes: &[u8; 64]) -> Self

Obtains a field element that is congruent to the provided little endian byte representation of an integer.

fn get_lower_128(&self) -> u128

Gets the lower 128 bits of this field element when expressed canonically.

fn get_lower_32(&self) -> u32

Gets the lower 32 bits of this field element when expressed canonically.

Provided methods

fn sqrt_alt(&self) -> (Choice, Self)

Equivalent to sqrt_ratio(self, one()).

fn rand() -> Self

This computes a random element of the field using system randomness.

fn write<W>(&self, writer: &mut W) -> Result<(), Error> where
W: Write,

Writes this element in its normalized, little endian form into a buffer.

fn read<R>(reader: &mut R) -> Result<Self, Error> where
R: Read,

Reads a normalized, little endian represented field element from a buffer.

fn pow(&self, by: &[u64; 4]) -> Self

Exponentiates self by by, where by is a little-endian order integer exponent.

fn pow_by_t_minus1_over2(&self) -> Self

Raise this field element to the power T_MINUS1_OVER2. Field implementations may override this to use an efficient addition chain.