Trait halo2::arithmetic::Group

pub trait Group: 'static + Copy + Clone + Send + Sync {
    type Scalar: FieldExt;
    fn group_zero() -> Self;

    fn group_add(&mut self, rhs: &Self);

    fn group_sub(&mut self, rhs: &Self);

    fn group_scale(&mut self, by: &Self::Scalar);
}

This represents an element of a group with basic operations that can be performed. This allows an FFT implementation (for example) to operate generically over either a field or elliptic curve group.

Associated Types

type Scalar: FieldExt

The group is assumed to be of prime order $p$. Scalar is the associated scalar field of size $p$.

Required methods

fn group_zero() -> Self

Returns the additive identity of the group.

fn group_add(&mut self, rhs: &Self)

Adds rhs to this group element.

fn group_sub(&mut self, rhs: &Self)

Subtracts rhs from this group element.

fn group_scale(&mut self, by: &Self::Scalar)

Scales this group element by a scalar.