Trait halo2::circuit::layouter::RegionLayouter

pub trait RegionLayouter<F: Field>: Debug {
    fn enable_selector<'v>(
        &'v mut self, 
        annotation: &'v (dyn Fn() -> String + 'v), 
        selector: &Selector, 
        offset: usize
    ) -> Result<(), Error>;

    fn assign_advice<'v>(
        &'v mut self, 
        annotation: &'v (dyn Fn() -> String + 'v), 
        column: Column<Advice>, 
        offset: usize, 
        to: &'v mut (dyn FnMut() -> Result<Assigned<F>, Error> + 'v)
    ) -> Result<Cell, Error>;

    fn assign_fixed<'v>(
        &'v mut self, 
        annotation: &'v (dyn Fn() -> String + 'v), 
        column: Column<Fixed>, 
        offset: usize, 
        to: &'v mut (dyn FnMut() -> Result<Assigned<F>, Error> + 'v)
    ) -> Result<Cell, Error>;

    fn constrain_equal(
        &mut self, 
        permutation: &Permutation, 
        left: Cell, 
        right: Cell
    ) -> Result<(), Error>;
}

Helper trait for implementing a custom Layouter.

This trait is used for implementing region assignments:

impl<'a, F: FieldExt, C: Chip<F>, CS: Assignment<F> + 'a> Layouter<C> for MyLayouter<'a, C, CS> {
    fn assign_region(
        &mut self,
        assignment: impl FnOnce(Region<'_, F, C>) -> Result<(), Error>,
    ) -> Result<(), Error> {
        let region_index = self.regions.len();
        self.regions.push(self.current_gate);

        let mut region = MyRegion::new(self, region_index);
        {
            let region: &mut dyn RegionLayouter<F> = &mut region;
            assignment(region.into())?;
        }
        self.current_gate += region.row_count;

        Ok(())
    }
}

TODO: It would be great if we could constrain the columns in these types to be “logical” columns that are guaranteed to correspond to the chip (and have come from Chip::Config).

Required methods

fn enable_selector<'v>(
&'v mut self,
annotation: &'v (dyn Fn() -> String + 'v),
selector: &Selector,
offset: usize
) -> Result<(), Error>

Enables a selector at the given offset.

fn assign_advice<'v>(
&'v mut self,
annotation: &'v (dyn Fn() -> String + 'v),
column: Column<Advice>,
offset: usize,
to: &'v mut (dyn FnMut() -> Result<Assigned<F>, Error> + 'v)
) -> Result<Cell, Error>

Assign an advice column value (witness)

fn assign_fixed<'v>(
&'v mut self,
annotation: &'v (dyn Fn() -> String + 'v),
column: Column<Fixed>,
offset: usize,
to: &'v mut (dyn FnMut() -> Result<Assigned<F>, Error> + 'v)
) -> Result<Cell, Error>

Assign a fixed value

fn constrain_equal(
&mut self,
permutation: &Permutation,
left: Cell,
right: Cell
) -> Result<(), Error>

Constraint two cells to have the same value.

Returns an error if either of the cells is not within the given permutation.