Function halo2::poly::commitment::verify_proof

pub fn verify_proof<'a, C: CurveAffine, E: EncodedChallenge<C>, T: TranscriptRead<C, E>>(
    params: &'a Params<C>, 
    msm: MSM<'a, C>, 
    transcript: &mut T, 
    x: C::Scalar, 
    v: C::Scalar
) -> Result<Guard<'a, C, E>, Error>

Checks to see if the proof represented within transcript is valid, and a point x that the polynomial commitment P opens purportedly to the value v. The provided msm should evaluate to the commitment P being opened.